Rename parse module to ir
