+++ savvy-bindgen/src/ir/savvy_enum.rs
