+++ savvy-bindgen/src/ir/savvy_struct.rs
