+++ savvy-bindgen/src/lib.rs
@@ -1,5 +1,5 @@
 mod gen;
-mod parse;
+mod ir;
 mod parse_file;
 mod utils;
 
@@ -9,11 +9,11 @@ pub use gen::static_files::{
     generate_cargo_toml, generate_config_toml, generate_configure, generate_example_lib_rs,
     generate_gitignore, generate_makevars_in, generate_makevars_win, generate_win_def,
 };
-pub use parse::savvy_enum::SavvyEnum;
-pub use parse::savvy_fn::{SavvyFn, SavvyFnArg, SavvyFnType};
-pub use parse::savvy_impl::SavvyImpl;
-pub use parse::savvy_struct::SavvyStruct;
-pub use parse::{merge_parsed_results, MergedResult, ParsedResult};
+pub use ir::savvy_enum::SavvyEnum;
+pub use ir::savvy_fn::{SavvyFn, SavvyFnArg, SavvyFnType};
+pub use ir::savvy_impl::SavvyImpl;
+pub use ir::savvy_struct::SavvyStruct;
+pub use ir::{merge_parsed_results, MergedResult, ParsedResult};
 
 pub use utils::extract_docs;
